feat: Написать алгоритм преобразования транзакции т-банка #35
- Сделал соотношение категорий
- Поиск категорий
- Создание корректных системных транзакций
- Исправление ошибок

+++ src/main/java/tech/nightsky/budgetly/tbankImport/TbankImportSummary.java
@@ -3,10 +3,10 @@ package tech.nightsky.budgetly.tbankImport;
 public record TbankImportSummary(
         Long id,
         Integer transactionCount,
-        Long skipped,
-        Long saved,
-        Long parsed,
-        Long filtered,
+        Integer skipped,
+        Integer saved,
+        Integer parsed,
+        Integer filtered,
         ImportStatus status
 ) {
 }
