+++ src/main/java/tech/nightsky/budgetly/tbankImport/api/TbankImportResponse.java
@@ -5,9 +5,9 @@ import tech.nightsky.budgetly.tbankImport.ImportStatus;
 public record TbankImportResponse(
         Long id,
         Integer transactionCount,
-        Long skipped,
-        Long saved,
-        Long parsed,
-        Long filtered,
+        Integer skipped,
+        Integer saved,
+        Integer parsed,
+        Integer filtered,
         ImportStatus status) {
 }
